Refactored scan filter manager to have only a default filter and one ephemeral filter Added options for how the ephemeral filter is applied
