+++ integration/src/androidTest/java/com/idevicesinc/sweetblue/ConnectBlitzer.java
@@ -34,7 +34,7 @@ public class ConnectBlitzer extends BaseTester<MainActivity>
     @Override BleManagerConfig getInitialConfig()
     {
         BleManagerConfig config = new BleManagerConfig();
-        config.defaultScanFilter = new BleManagerConfig.ScanFilter()
+        config.defaultScanFilter = new ScanFilter()
         {
             @Override public Please onEvent(ScanEvent e)
             {
